phoenix_integration/requirements.txt
# Phoenix Evaluation Framework
arize-phoenix[evals]>=4.0.0

# LLM Providers
openai>=1.0.0
groq>=0.4.0
anthropic>=0.18.0

# Database
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Async support
asyncio-throttle>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
